修复部分情况下获取登录账号UID为-1的问题 指令增加关键词匹配

+++ Sora/OnebotInterface/ApiInterface.cs
@@ -119,7 +119,7 @@ namespace Sora.OnebotInterface
             return
             (
                 retCode,
-                uid: int.TryParse(ret?["data"]?["user_id"]?.ToString(), out var uid) ? uid : -1,
+                uid: long.TryParse(ret?["data"]?["user_id"]?.ToString(), out var uid) ? uid : -1,
                 nick: ret?["data"]?["nickname"]?.ToString() ?? string.Empty
             );
         }
